Conflicts:
	main/anagrafe.py

+++ main/anagrafe.py
@@ -14,6 +14,7 @@ from main import data_render
 from main import database_manager
 from main import scripts
 from main import views
+from main import users
 
 import re
 import logging
@@ -40,7 +41,7 @@ def __breadcrumbName(l):
 
     return s
 
-def show_record(request, cliente_id, detail_type=None, impianto_id=None, sub_impianto_id=None, hdr_message='', message=''):
+def show_record(request, cliente_id, detail_type=None, impianto_id=None, sub_impianto_id=None, message_hdr='', message='', message_type=''):
     data, data_list = view_record(cliente_id, detail_type, impianto_id, sub_impianto_id)
 
 
@@ -49,25 +50,24 @@ def show_record(request, cliente_id, detail_type=None, impianto_id=None, sub_imp
     r = re.compile("\w+\/\d+")
     #['anagrafe/33', 'impianto/33', 'verifica/33']
     l = r.findall(request.path)
-    url = "/"
-    if len(l) > 1:
-        print "qui..", url
-        for i in l[:-1]:
-            url += i + "/"
-            s = __breadcrumbName(i)
-            b += BREADCRUMB_ELEMENT % (url, s, BREADCRUMB_DIVIDER)
+    if l:
+        url = "/"
+        if len(l) > 1:
+            for i in l[:-1]:
+                url += i + "/"
+                s = __breadcrumbName(i)
+                b += BREADCRUMB_ELEMENT % (url, s, BREADCRUMB_DIVIDER)
 
-    url += l[-1] + "/"
-    s = __breadcrumbName(l[-1])
-    b += BREADCRUMB_ELEMENT % (url, s, "")
+        url += l[-1] + "/"
+        s = __breadcrumbName(l[-1])
+        b += BREADCRUMB_ELEMENT % (url, s, "")
 
     if data is None:
         _display_error(request, "Qualcosa e' andato storto!")
 
     return render(request, 'anagrafe.sub', { 'data': data,
                            'data_list': data_list,
-                           'notification_hdr': hdr_message,
-                           'notification_msg': message,
+                           'notification': data_render.notification(message_hdr, message, message_type),
                            'scripts':scripts.ANAGRAFE_JS,
                            'breadcrumb': b,
                            })
@@ -251,16 +251,40 @@ def view_record(cliente_id, detail_type=None, impianto_id=None, sub_impianto_id=
     return data, data_list
 
 
-ERROR_CLIENTE_INSERITO = """<div class=\"alert alert-block\"> <button type=\"button\" class=\"close\" data-dismiss=\"alert\">&times;</button>
-<h4>Attenzione!</h4>
-Cliente gia\' inserito nel gestionale..
-</div>"""
+ERROR_DB_HDR = "Database Error."
+ERROR_DB_CLIENTE_INSERITO = "Cliente gia\' inserito nel gestionale."
+ERROR_DB_DELETE = "Record gia\' rimosso"
 
-ERROR_FORM = """<div class=\"alert alert-error\">
-<button type=\"button\" class=\"close\" data-dismiss=\"alert\">&times;</button>
-<h4>Errore!</h4>
-Errore nel form..
-</div>"""
+ERROR_FORM_HDR = "Errore form."
+ERROR_FORM = "Ci sono errori nella compilazione del form."
+
+ERROR_PERM_HDR = "Errore di permesso."
+ERROR_PERM_ADD = "Non hai il permesso di Aggiungere record al gestionale."
+ERROR_PERM_EDIT = "Non hai il permesso di Modificare il record del gestionale."
+ERROR_PERM_DELETE = "Non hai il permesso di Cancellare il record dal gestionale."
+
+OK_DB_DELETE = "Cancellazione riuscita."
+
+
+def __checkPermission(request, select, form, d):
+    if select is None:
+        if users.has_addPermission(request, select):
+            return True, d
+
+        d['message'] = ERROR_PERM_ADD
+
+    else:
+        if users.has_editPermission(request, select):
+            return True, d
+
+        d['message'] = ERROR_PERM_EDIT
+
+
+    d['message_hdr'] = ERROR_PERM_HDR
+    d['message_type'] = data_render.MSG_ERROR
+    d['form'] = form
+
+    return False, d
 
 def __editAdd_record(cliente_id, impianto_id, sub_impianto_id, detail_type, request, select=None):
     """
@@ -273,17 +297,25 @@ def __editAdd_record(cliente_id, impianto_id, sub_impianto_id, detail_type, requ
         'cliente_id':None,
         'impianto_id':None,
         'sub_impianto_id':None,
-        'message':None,
+        'message_hdr':'',
+        'message': '',
+        'message_type': '',
     }
 
     if cliente_id is None or detail_type is None:
         msg = None
         form = models.ClienteForm(request.POST, instance=select)
+        ret, d = __checkPermission(request, select, form, d)
+        if not ret:
+            return False, d
+
         if form.is_valid():
             cli = form.cleaned_data['cliente_id_inserito']
             if cli is not None:
                 d['cliente_id'] = cli
-                d['message'] = ERROR_CLIENTE_INSERITO
+                d['message_hdr'] = ERROR_DB_HDR
+                d['message'] = ERROR_DB_CLIENTE_INSERITO
+                d['message_type'] = data_render.MSG_ERROR
                 return True, d
 
             instance = form.save()
@@ -292,24 +324,37 @@ def __editAdd_record(cliente_id, impianto_id, sub_impianto_id, detail_type, requ
             return True, d
 
         else:
+            d['message'] = ERROR_FORM
+            d['message_hdr'] = ERROR_FORM_HDR
+            d['message_type'] = data_render.MSG_ERROR
             d['form'] = form
             return False, d
-
     else:
         if detail_type == 'impianto':
             form = models.ImpiantoForm(request.POST, instance=select)
+            ret, d = __checkPermission(request, select, form, d)
+            if not ret:
+                return False, d
+
             if form.is_valid():
                 instance = form.save()
                 d['cliente_id'] = cliente_id
                 d['impianto_id'] = instance.id
                 return True, d
             else:
+                d['form'] = form
                 d['message'] = ERROR_FORM
+                d['message_hdr'] = ERROR_FORM_HDR
+                d['message_type'] = data_render.MSG_ERROR
                 d['form'] = form
                 return False, d
 
         if detail_type == 'verifica':
             form = models.VerificaForm(request.POST, instance=select)
+            ret, d = __checkPermission(request, select, form, d)
+            if not ret:
+                return False, d
+
             if form.is_valid():
                 instance = form.save()
                 d['cliente_id'] = cliente_id
@@ -320,10 +365,16 @@ def __editAdd_record(cliente_id, impianto_id, sub_impianto_id, detail_type, requ
             else:
                 d['form'] = form
                 d['message'] = ERROR_FORM
+                d['message_hdr'] = ERROR_FORM_HDR
+                d['message_type'] = data_render.MSG_ERROR
                 return False, d
 
         if detail_type == 'intervento':
             form = models.InterventoForm(request.POST, instance=select)
+            ret, d = __checkPermission(request, select, form, d)
+            if not ret:
+                return False, d
+
             if form.is_valid():
                 instance = form.save(commit = False)
                 instance.tipo_intervento = instance.tipo_intervento.capitalize()
@@ -334,8 +385,10 @@ def __editAdd_record(cliente_id, impianto_id, sub_impianto_id, detail_type, requ
                 d['detail_type'] = detail_type
                 return True, d
             else:
-                d['message'] = ERROR_FORM
                 d['form'] = form
+                d['message'] = ERROR_FORM
+                d['message_hdr'] = ERROR_FORM_HDR
+                d['message_type'] = data_render.MSG_ERROR
                 return False, d
 
     return _display_error(request, "Qualcosa e' andato storto..")
@@ -344,6 +397,7 @@ def add_record(request, cliente_id=None, detail_type=None, impianto_id=None, sub
     script = ""
     data = ""
     data_list = ""
+    notification = ""
     if cliente_id is None:
         header_msg = "Aggiungi Nuovo Cliente"
         post_url = "add/"
@@ -389,9 +443,10 @@ def add_record(request, cliente_id=None, detail_type=None, impianto_id=None, sub
         else:
             request = d['request']
             form = d['form']
+            notification = data_render.notification(d['message_hdr'], d['message'], d['message_type'])
 
-    return render(request, 'anagrafe_form.sub', {'header_msg': data_render.TITLE_STYLE_FORM % header_msg, 'data_forms': form,
-        'data':data, 'scripts': script, 'post_url':post_url})
+    return render(request, 'anagrafe_form.sub', {'notification': notification, 'header_msg': data_render.TITLE_STYLE_FORM % header_msg,
+        'data_forms': form, 'data':data, 'scripts': script, 'post_url':post_url})
 
 
 def edit_record(request, cliente_id=None, detail_type=None, impianto_id=None, sub_impianto_id=None):
@@ -402,6 +457,7 @@ def edit_record(request, cliente_id=None, detail_type=None, impianto_id=None, su
     data = ""
     data_list = ""
     script = ""
+    notification = ""
     if detail_type is None:
         select = models.Cliente.objects.get(pk=cliente_id)
         form = models.ClienteForm(instance=select)
@@ -440,9 +496,10 @@ def edit_record(request, cliente_id=None, detail_type=None, impianto_id=None, su
         else:
             request = d['request']
             form = d['form']
+            notification = data_render.notification(d['message_hdr'], d['message'], d['message_type'])
 
-    return render(request, 'anagrafe_form.sub', {'header_msg': data_render.TITLE_STYLE_FORM % header_msg, 'data_forms': form,
-        'data':data, 'scripts': script, 'post_url':post_url})
+    return render(request, 'anagrafe_form.sub', {'notification': notification, 'header_msg': data_render.TITLE_STYLE_FORM % header_msg,
+        'data_forms': form, 'data':data, 'scripts': script, 'post_url':post_url})
 
 def delete_record(request, cliente_id=None, detail_type=None, impianto_id=None, sub_impianto_id=None):
     data = ''
@@ -450,14 +507,23 @@ def delete_record(request, cliente_id=None, detail_type=None, impianto_id=None, 
     data_list = ''
     try:
         if request.method == 'POST':
+            if not users.has_deletePermission(request):
+                return show_record(request, cliente_id=cliente_id, detail_type=detail_type, impianto_id=impianto_id, sub_impianto_id=sub_impianto_id,
+                        message_hdr=ERROR_PERM_HDR, message=ERROR_PERM_DELETE, message_type=data_render.MSG_ERROR)
+
             if detail_type is None:
                 cli = models.Cliente.objects.get(pk=cliente_id)
                 nome = cli.nome
                 cognome = cli.cognome
                 cli.delete()
                 s = "Cliente: %s %s Rimosso correttamente." % (nome, cognome)
+                d = {
+                    'message_hdr': OK_DB_DELETE ,
+                    'message': s,
+                    'message_type':data_render.MSG_OK,
+                 }
 
-                return views.home(request, s)
+                return views.home(request, d)
 
             else:
                 if detail_type == 'impianto':
@@ -465,24 +531,28 @@ def delete_record(request, cliente_id=None, detail_type=None, impianto_id=None, 
                     s = "%s" % imp
                     imp.delete()
                     s = "Impianto: %s rimosso correttamente." % s
-                    return show_record(request, cliente_id=cliente_id, hdr_message="Ok!", message=s)
+                    return show_record(request, cliente_id=cliente_id,
+                            message_hdr=OK_DB_DELETE, message=s, message_type=data_render.MSG_OK)
 
                 if detail_type == 'verifica':
                     ver = models.Verifica.objects.get(pk=sub_impianto_id)
                     s = "%s" % ver
                     ver.delete()
                     s = "Verifica e Manutezione: %s rimosso correttamente." % s
-                    return show_record(request, cliente_id=cliente_id, impianto_id=impianto_id, hdr_message="Ok!", message=s)
+                    return show_record(request, cliente_id=cliente_id, impianto_id=impianto_id,
+                            message_hdr=OK_DB_DELETE, message=s, message_type=data_render.MSG_OK)
 
                 if detail_type == 'intervento':
                     interv = models.Intervento.objects.get(pk=sub_impianto_id)
                     s = "%s" % interv
                     interv.delete()
                     s = "Verifica e Manutezione: %s rimosso correttamente." % s
-                    return show_record(request, cliente_id=cliente_id, impianto_id=impianto_id, hdr_message="Ok!", message=s)
+                    return show_record(request, cliente_id=cliente_id, impianto_id=impianto_id,
+                            message_hdr=OK_DB_DELETE, message=s, message_type=data_render.MSG_OK)
 
     except ObjectDoesNotExist, m:
-        return _display_error(request, "Qualcosa e' andato storto..(%s)" % m)
+        return show_record(request, cliente_id=cliente_id, detail_type=detail_type, impianto_id=impianto_id, sub_impianto_id=sub_impianto_id,
+                message_hdr=ERROR_DB_HDR, message=ERROR_DB_DELETE, message_type=data_render.MSG_ERROR)
 
     return _display_error(request, "Qualcosa e' andato storto..")
